extend to new test scenario
