+++ xios_examples/read_domain_resample/__init__.py
@@ -1,0 +1,3 @@
+"""
+Enable this folder to be a module path, for imports and test discovery.
+"""
