+++ xios_examples/read_domain_resample/xios.xml
@@ -1,0 +1,22 @@
+<context>
+  <variable_definition>
+    <variable_group id="buffer">
+      <variable id="optimal_buffer_size" type="string">
+	performance
+      </variable>
+      <variable id="buffer_size_factor" type="double">
+	1.0
+      </variable>
+    </variable_group>
+    
+    <variable_group id="parameters" >
+       <variable id="using_server" type="bool">true</variable>
+       <variable id="info_level" type="int">
+	100
+      </variable>
+      <variable id="print_file" type="bool">
+	true
+      </variable>
+    </variable_group>
+  </variable_definition>
+</context>
